third_party/blink/web_tests/editing/text-iterator/range-to-from-location-and-length-expected.txt

PASS locationAndLengthFromRange(range(p.firstChild, 0, p.firstChild, 3)) is [0, 1]
PASS locationAndLengthFromRange(range(p.firstChild, 1, p.firstChild, 4)) is [0, 2]
PASS locationAndLengthFromRange(range(p.firstChild, 2, p.firstChild, 3)) is [0, 1]
PASS internals.rangeFromLocationAndLength(container, 0, 1).toArray() is [p.firstChild, 2, p.firstChild, 3]
PASS locationAndLengthFromRange(range(p.firstChild, 2, p.firstChild, 7)) is [0, 5]
PASS locationAndLengthFromRange(range(p.firstChild, 2, p, 1)) is [0, 5]
PASS locationAndLengthFromRange(range(p.firstChild, 2, p, 2)) is [0, 6]
PASS locationAndLengthFromRange(range(p.firstChild, 2, p.childNodes[2], 1)) is [0, 6]
PASS internals.rangeFromLocationAndLength(container, 1, 4).toArray() is [p.firstChild, 3, p.firstChild, 7]
PASS internals.rangeFromLocationAndLength(container, 0, 6).toArray() is [p.firstChild, 2, p.childNodes[2], 1]
PASS locationAndLengthFromRange(range(p.childNodes[2], 6, p, 3)) is [11, 1]
FAIL internals.rangeFromLocationAndLength(container, 11, 2).toArray() should be [object Text],6,[object HTMLDivElement],0. Was [object Text],6,[object HTMLParagraphElement],3.
PASS successfullyParsed is true

TEST COMPLETE
